requirements.txt
streamlit
requests
Pillow
torch
transformers
google-generativeai
diffusers